back-end/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
#openai
google-genai
pytest
httpx
jsonschema
pytest-asyncio
pydantic
python-multipart
supabase
requests
psycopg2-binary
sqlalchemy[postgresql]